requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
jinja2
google-cloud-secret-manager